+++ Changelog.md
@@ -6,6 +6,8 @@
   * Pass `Document` instance to transformation callbacks (#86)
   * Made `nokogiri` dependency more forgiving.
     * Supports (1.5.0)...(1.7.0) now instead of (1.6.0)...(1.7.0). Some people out there are stuck on this older version of Nokogiri, and I don't want to leave them out.
+  * Output better errors when no assets can be found.
+    * The error will now show which providers were tried and in which order, along with the error message from the specific providers.
 
 ### 3.0.5